# Create-Attachments/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MirrorJobInput {
  /** HTTPS or HTTP URL for a site the caller owns or is authorized to archive */
  url: string;
  /**
     * @minimum 1
     * @maximum 1000
     */
  maxPages?: number;
  /**
     * @minimum 0
     * @maximum 5000
     */
  requestDelayMs?: number;
  respectRobotsTxt?: boolean;
  /**
     * @minimum 0
     * @maximum 10
     */
  maxDepth?: number;
  includeAssets?: boolean;
  /**
     * @minLength 1
     * @maxLength 200
     */
  pathPrefix?: string;
  /**
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 200
     */
  excludePaths?: string[];
  /**
     * Wall-clock limit for the whole crawl, in milliseconds.
     * @minimum 60000
     * @maximum 3600000
     */
  timeoutMs?: number;
  /**
     * Overall archive size cap, in bytes.
     * @minimum 1048576
     * @maximum 2147483648
     */
  maxTotalBytes?: number;
}

export type MirrorJobStatus = typeof MirrorJobStatus[keyof typeof MirrorJobStatus];


export const MirrorJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  completed_with_warnings: 'completed_with_warnings',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type MirrorJobProgressPhase = typeof MirrorJobProgressPhase[keyof typeof MirrorJobProgressPhase];


export const MirrorJobProgressPhase = {
  queued: 'queued',
  discovering: 'discovering',
  saving: 'saving',
  downloading_assets: 'downloading_assets',
  rewriting: 'rewriting',
  packaging: 'packaging',
} as const;

export interface MirrorQueueSummary {
  saved: number;
  skipped: number;
  failed: number;
  discovered: number;
  pending: number;
}

export type MirrorOutcomeKind = typeof MirrorOutcomeKind[keyof typeof MirrorOutcomeKind];


export const MirrorOutcomeKind = {
  page: 'page',
  asset: 'asset',
} as const;

export type MirrorOutcomeStatus = typeof MirrorOutcomeStatus[keyof typeof MirrorOutcomeStatus];


export const MirrorOutcomeStatus = {
  saved: 'saved',
  skipped: 'skipped',
  failed: 'failed',
} as const;

export interface MirrorOutcome {
  kind: MirrorOutcomeKind;
  url: string;
  status: MirrorOutcomeStatus;
  /** @nullable */
  httpStatus: number | null;
  /** @nullable */
  contentType: string | null;
  /** @nullable */
  finalUrl: string | null;
  /** @nullable */
  archivePath: string | null;
  /** @nullable */
  reason: string | null;
  attempts: number;
  bytes: number;
  updatedAt: string;
}

export interface MirrorJob {
  id: string;
  url: string;
  status: MirrorJobStatus;
  pagesFound: number;
  pagesDownloaded: number;
  pagesSkipped: number;
  pagesFailed: number;
  assetsDownloaded: number;
  assetsSkipped: number;
  assetsFailed: number;
  bytesDownloaded: number;
  maxPages: number;
  requestDelayMs: number;
  respectRobotsTxt: boolean;
  maxDepth?: number;
  includeAssets?: boolean;
  pathPrefix?: string;
  excludePaths?: string[];
  timeoutMs?: number;
  maxTotalBytes?: number;
  /** @nullable */
  currentUrl: string | null;
  progressPhase: MirrorJobProgressPhase;
  /** @nullable */
  message: string | null;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
  archiveAvailable: boolean;
  throughputBytesPerSecond: number;
  estimatedRemainingMs: number | null;
  warningCount: number;
  errorCount: number;
  resumable: boolean;
  events: Array<{ at: string; type: string; message: string; url?: string | null }>;
  queueSummary: MirrorQueueSummary;
  outcomes: MirrorOutcome[];
}

export interface MirrorJobList {
  jobs: MirrorJob[];
}

export type MirrorArchiveFileKind = typeof MirrorArchiveFileKind[keyof typeof MirrorArchiveFileKind];


export const MirrorArchiveFileKind = {
  page: 'page',
  asset: 'asset',
} as const;

export type MirrorArchiveFileStatus = typeof MirrorArchiveFileStatus[keyof typeof MirrorArchiveFileStatus];


export const MirrorArchiveFileStatus = {
  saved: 'saved',
  skipped: 'skipped',
  failed: 'failed',
} as const;

export interface MirrorArchiveFile {
  path: string;
  kind: MirrorArchiveFileKind;
  url: string;
  status: MirrorArchiveFileStatus;
  /** @nullable */
  contentType: string | null;
  bytes: number;
  /** @nullable */
  reason: string | null;
  /** @nullable */
  finalUrl: string | null;
}

export type MirrorArchiveManifestConfiguration = { [key: string]: unknown };

export type MirrorArchiveManifestSummary = { [key: string]: unknown };

export type MirrorArchiveManifestRedirects = {[key: string]: string};

export interface MirrorArchiveManifest {
  schemaVersion: number;
  jobId: string;
  sourceUrl: string;
  configuration: MirrorArchiveManifestConfiguration;
  summary: MirrorArchiveManifestSummary;
  redirects: MirrorArchiveManifestRedirects;
  outcomes: MirrorOutcome[];
}

export interface MirrorIntegrityWarning {
  kind: string;
  url: string;
  /** @nullable */
  path: string | null;
  status: string;
  /** @nullable */
  reason: string | null;
  /** @nullable */
  httpStatus: number | null;
}

export interface MirrorIntegrityReport {
  valid: boolean;
  schemaVersion: number;
  fileCount: number;
  savedCount: number;
  warningCount: number;
  brokenPages: number;
  brokenAssets: number;
  warnings: MirrorIntegrityWarning[];
}

export interface ErrorResponse {
  error: string;
}

export type ListMirrorJobsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListMirrorArchiveFilesParams = {
search?: string;
kind?: ListMirrorArchiveFilesKind;
status?: ListMirrorArchiveFilesStatus;
};

export type ListMirrorArchiveFilesKind = typeof ListMirrorArchiveFilesKind[keyof typeof ListMirrorArchiveFilesKind];


export const ListMirrorArchiveFilesKind = {
  page: 'page',
  asset: 'asset',
} as const;

export type ListMirrorArchiveFilesStatus = typeof ListMirrorArchiveFilesStatus[keyof typeof ListMirrorArchiveFilesStatus];


export const ListMirrorArchiveFilesStatus = {
  saved: 'saved',
  skipped: 'skipped',
  failed: 'failed',
} as const;

export type ListMirrorArchiveFiles200 = {
  total: number;
  files: MirrorArchiveFile[];
};