# Create-Attachments/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: mirror
    description: Authorized website mirror jobs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /mirror-jobs:
    get:
      operationId: listMirrorJobs
      tags: [mirror]
      summary: List recent mirror jobs
      description: Returns the most recently created mirror jobs, newest first.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            multipleOf: 1
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Recent mirror jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorJobList"
    post:
      operationId: createMirrorJob
      tags: [mirror]
      summary: Start an authorized website mirror job
      description: Begins a same-origin crawl for a website the caller owns or is authorized to archive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MirrorJobInput"
      responses:
        "202":
          description: Mirror job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorJob"
        "400":
          description: Invalid URL or crawl settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}:
    get:
      operationId: getMirrorJob
      tags: [mirror]
      summary: Get mirror job progress
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "200":
          description: Mirror job progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorJob"
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/cancel:
    post:
      operationId: cancelMirrorJob
      tags: [mirror]
      summary: Cancel a queued or running mirror job
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "200":
          description: Cancelled mirror job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorJob"
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/retry:
    post:
      operationId: retryMirrorJob
      tags: [mirror]
      summary: Start another run with the same mirror settings
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "202":
          description: Retry mirror job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorJob"
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mirror is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/resume:
    post:
      operationId: resumeMirrorJob
      tags: [mirror]
      summary: Resume unfinished work for a cancelled or failed mirror
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "202":
          description: Resume accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorJob"
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mirror cannot be resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/events:
    get:
      operationId: streamMirrorJobEvents
      tags: [mirror]
      summary: Stream live mirror progress snapshots
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "200":
          description: Server-sent progress snapshots
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/archive/manifest:
    get:
      operationId: getMirrorArchiveManifest
      tags: [mirror]
      summary: Read a sealed mirror archive manifest
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "200":
          description: Archive manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorArchiveManifest"
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Archive is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/archive/files:
    get:
      operationId: listMirrorArchiveFiles
      tags: [mirror]
      summary: Search files in a sealed mirror archive
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
        - name: search
          in: query
          schema:
            type: string
        - name: kind
          in: query
          schema:
            type: string
            enum: [page, asset]
        - name: status
          in: query
          schema:
            type: string
            enum: [saved, skipped, failed]
      responses:
        "200":
          description: Matching archive files
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
                  files:
                    type: array
                    items:
                      $ref: "#/components/schemas/MirrorArchiveFile"
                required: [total, files]
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Archive is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/archive/integrity:
    get:
      operationId: getMirrorArchiveIntegrity
      tags: [mirror]
      summary: Inspect archive health and failed resources
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "200":
          description: Archive integrity report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MirrorIntegrityReport"
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Archive is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/download:
    get:
      operationId: downloadMirrorJob
      tags: [mirror]
      summary: Download a completed mirror as a ZIP archive
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "200":
          description: ZIP archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: Mirror job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mirror is not complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/preview:
    get:
      operationId: previewMirrorJob
      tags: [mirror]
      summary: Open the starting page of a completed mirror
      description: Redirects to the archived path for the mirrored starting page.
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
      responses:
        "302":
          description: Redirect to the archived starting page
          headers:
            Location:
              schema:
                type: string
        "404":
          description: Mirror job or starting page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mirror is not complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mirror-jobs/{id}/preview/{previewPath}:
    get:
      operationId: serveMirrorPreviewFile
      tags: [mirror]
      summary: Serve a file from a completed mirror preview
      parameters:
        - $ref: "#/components/parameters/MirrorJobId"
        - name: previewPath
          in: path
          required: true
          description: Safe slash-separated path inside the mirror archive
          schema:
            type: string
      responses:
        "200":
          description: Mirrored file
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Mirror job or preview file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mirror is not complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    MirrorJobId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MirrorJobInput:
      type: object
      properties:
        url:
          type: string
          description: HTTPS or HTTP URL for a site the caller owns or is authorized to archive
        maxPages:
          type: number
          multipleOf: 1
          minimum: 1
          maximum: 1000
          default: 100
        requestDelayMs:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 5000
          default: 250
        respectRobotsTxt:
          type: boolean
          default: true
        maxDepth:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 10
          default: 3
        includeAssets:
          type: boolean
          default: true
        pathPrefix:
          type: string
          minLength: 1
          maxLength: 200
          default: /
        excludePaths:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 200
        timeoutMs:
          type: number
          multipleOf: 1
          minimum: 60000
          maximum: 3600000
          default: 900000
          description: Wall-clock limit for the whole crawl, in milliseconds.
        maxTotalBytes:
          type: number
          multipleOf: 1
          minimum: 1048576
          maximum: 2147483648
          default: 524288000
          description: Overall archive size cap, in bytes.
      required:
        - url
    MirrorJobList:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/MirrorJob"
      required:
        - jobs
    MirrorJob:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [queued, running, completed, completed_with_warnings, failed, cancelled]
        pagesFound:
          type: number
          multipleOf: 1
        pagesDownloaded:
          type: number
          multipleOf: 1
        pagesSkipped:
          type: number
          multipleOf: 1
        pagesFailed:
          type: number
          multipleOf: 1
        assetsDownloaded:
          type: number
          multipleOf: 1
        assetsSkipped:
          type: number
          multipleOf: 1
        assetsFailed:
          type: number
          multipleOf: 1
        bytesDownloaded:
          type: number
          multipleOf: 1
        maxPages:
          type: number
          multipleOf: 1
        requestDelayMs:
          type: number
          multipleOf: 1
        respectRobotsTxt:
          type: boolean
        maxDepth:
          type: number
          multipleOf: 1
        includeAssets:
          type: boolean
        pathPrefix:
          type: string
        excludePaths:
          type: array
          items:
            type: string
        timeoutMs:
          type: number
          multipleOf: 1
        maxTotalBytes:
          type: number
          multipleOf: 1
        currentUrl:
          type: ["string", "null"]
        progressPhase:
          type: string
          enum: [queued, discovering, saving, downloading_assets, rewriting, packaging]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
        archiveAvailable:
          type: boolean
        throughputBytesPerSecond:
          type: number
        estimatedRemainingMs:
          type: ["number", "null"]
        warningCount:
          type: number
        errorCount:
          type: number
        resumable:
          type: boolean
        events:
          type: array
          items:
            type: object
            properties:
              at:
                type: string
                format: date-time
              type:
                type: string
              message:
                type: string
              url:
                type: ["string", "null"]
            required: [at, type, message]
        queueSummary:
          $ref: "#/components/schemas/MirrorQueueSummary"
        outcomes:
          type: array
          items:
            $ref: "#/components/schemas/MirrorOutcome"
      required:
        - id
        - url
        - status
        - pagesFound
        - pagesDownloaded
        - pagesSkipped
        - pagesFailed
        - assetsDownloaded
        - assetsSkipped
        - assetsFailed
        - bytesDownloaded
        - maxPages
        - requestDelayMs
        - respectRobotsTxt
        - currentUrl
        - progressPhase
        - message
        - createdAt
        - completedAt
        - archiveAvailable
        - throughputBytesPerSecond
        - estimatedRemainingMs
        - warningCount
        - errorCount
        - resumable
        - events
        - queueSummary
        - outcomes
    MirrorQueueSummary:
      type: object
      properties:
        saved:
          type: number
        skipped:
          type: number
        failed:
          type: number
        discovered:
          type: number
        pending:
          type: number
      required: [saved, skipped, failed, discovered, pending]
    MirrorOutcome:
      type: object
      properties:
        kind:
          type: string
          enum: [page, asset]
        url:
          type: string
        status:
          type: string
          enum: [saved, skipped, failed]
        httpStatus:
          type: ["number", "null"]
        contentType:
          type: ["string", "null"]
        finalUrl:
          type: ["string", "null"]
        archivePath:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        attempts:
          type: number
        bytes:
          type: number
        updatedAt:
          type: string
          format: date-time
      required: [kind, url, status, httpStatus, contentType, finalUrl, archivePath, reason, attempts, bytes, updatedAt]
    MirrorArchiveFile:
      type: object
      properties:
        path:
          type: string
        kind:
          type: string
          enum: [page, asset]
        url:
          type: string
        status:
          type: string
          enum: [saved, skipped, failed]
        contentType:
          type: ["string", "null"]
        bytes:
          type: number
        reason:
          type: ["string", "null"]
        finalUrl:
          type: ["string", "null"]
      required: [path, kind, url, status, contentType, bytes, reason, finalUrl]
    MirrorArchiveManifest:
      type: object
      properties:
        schemaVersion:
          type: number
        jobId:
          type: string
        sourceUrl:
          type: string
        configuration:
          type: object
          additionalProperties: true
        summary:
          type: object
          additionalProperties: true
        redirects:
          type: object
          additionalProperties:
            type: string
        outcomes:
          type: array
          items:
            $ref: "#/components/schemas/MirrorOutcome"
      required: [schemaVersion, jobId, sourceUrl, configuration, summary, redirects, outcomes]
    MirrorIntegrityWarning:
      type: object
      properties:
        kind:
          type: string
        url:
          type: string
        path:
          type: ["string", "null"]
        status:
          type: string
        reason:
          type: ["string", "null"]
        httpStatus:
          type: ["number", "null"]
      required: [kind, url, path, status, reason, httpStatus]
    MirrorIntegrityReport:
      type: object
      properties:
        valid:
          type: boolean
        schemaVersion:
          type: number
        fileCount:
          type: number
        savedCount:
          type: number
        warningCount:
          type: number
        brokenPages:
          type: number
        brokenAssets:
          type: number
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/MirrorIntegrityWarning"
      required: [valid, schemaVersion, fileCount, savedCount, warningCount, brokenPages, brokenAssets, warnings]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error