# Create-Attachments/lib/api-zod/src/generated/types/mirrorJobList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MirrorJob } from './mirrorJob';

export interface MirrorJobList {
  jobs: MirrorJob[];
}